refactor: remove base_url arguments to get_mgmt_svc_client
The default value in `get_mgmt_svc_client` is the same as all these instances
of passing this argument into the function.  The parameter still
exists in the definition if somewhere else a different value is still
provided.

Eliminating this reduces a bunch of lines of code and reduces the
number of references to search through.  Especially since a lot of
these references are in child classes referencing a semi-private
variable (self._cloud_environment).