refactor: remove unused imports
These VERSION imports aren't referenced anywhere and cause problems
when trying to update the underlying azure libraries.

These aren't used to probe the presence of the library either as the
parent module is imported a few lines down. E.g. `azure.mgmt.monitor` is
imported so we don't need to the unused import of `azure.mgmt.monitor.version.VERSION`