refactor: Upgrade azure-common & azure-cli-core
The newer version of azure-common has a safely guarded import of
azure-cli-core. That logic doesn't need to be repeated here so remove
it for simplicity. This helps prepare for future work of replacing all
of this with `azure-identity`.

If azure-cli-core isn't installed and:

* `auth_source=cli` is specified: "msg: Azure CLI profile cannot be
loaded - You need to install 'azure-cli-core' to load CLI credentials"

* `auth_source=auto` is specified: "msg: Failed to get
credentials. Either pass as parameters, setenvironment variables,
define a profile in ~/.azure/credentials, or log in with Azure
CLI (`az login`)."